Making the readme instructions dummy(me) proof
I wanted to make some more explicit instructions that assume you don't
know anything about this sytem. This will allow us to rebuidl these
kiosks in a few years when we've forgotten how this whole system works.

This is a first pass before I checked in with Sarah, the primary author
of this this template system. The readme may need some editing.